+++ marxs/optics/multiLayerMirror.py
@@ -22,7 +22,8 @@ class FlatBrewsterMirror(FlatOpticalElement):
     probability of the reflected photons is adjusted to account for this overall loss.
     '''
     display = {'color': (0., 1., 0.),
-               'shape': 'box'
+               'shape': 'box',
+               'box-half': '+x',
     }
 
     def fresnel(self, photons, intersect, intersection, local):