+++ marxs/visualization/mayavi.py
@@ -80,13 +80,37 @@ def surface(surface, display, viewer=None):
 
 @format_doc(doc_plot)
 def box(obj, display, viewer=None):
-    '''Plot a rectangular box for an object.'''
+    '''Plot a rectangular box for an object.
+
+    By default, the box extends in x,y, and z direction. The display keyword
+    "box-half" can be used to show only one half of the box, e.g. "+x" would
+    show  the full extend in y and z direction, but only the lower half in
+    x direction (such that rays coming from the +x direction are
+    visible up to the interaction point).
+    Use this for elements such as mirrors or detectors where
+    photon interaction happens on the surface, not in the substrate.
+    '''
     from mayavi import mlab
 
     corners = np.array([[-1, -1, -1], [-1,+1, -1],
                         [-1, -1,  1], [-1, 1,  1],
                         [ 1, -1, -1], [ 1, 1, -1],
                         [ 1, -1, +1], [ 1, 1, +1]])
+    if 'box-half' in display:
+        # write in a way that it works with any value for that keyword
+        try:
+            if display['box-half'][0] == '+':
+                factor = +1
+            elif display['box-half'][0] == '-':
+                factor = -1
+            else:
+                factor = 0
+            xyz = {'x': 0, 'y': 1, 'z': 2}
+            if display['box-half'][1] in xyz:
+                j = xyz[display['box-half'][1]]
+                corners[corners[:, j] == factor, j] = 0
+        except:
+            pass
     triangles = [(0,2,6), (0,4,6), (0,1,5), (0,4,5), (0,1,3), (0,2,3),
                  (7,3,2), (7,6,2), (7,3,1), (7,5,1), (7,6,4), (7,5,4)]
     corners = np.einsum('ij,...j->...i', obj.pos4d, mutils.e2h(corners, 1))
@@ -96,7 +120,13 @@ def box(obj, display, viewer=None):
 
 @format_doc(doc_plot)
 def cylinder(obj, display, viewer=None):
-    '''Plot a rectangular box for an object.'''
+    '''Plot a cylinder for an object.
+
+    The radius is taken from the y dimensions, the length of the tube from
+    the x dimension.
+    The display keyword "tube-sides" (default: 20) sets the number of vortices
+    that are used to approximate the curve.
+    '''
     from mayavi import mlab
 
     x0 = obj.geometry('center') - obj.geometry('v_x')